drm_hwcomposer: hwcutils: add cast (int) to building (Android P)
Add cast to (int) to avoid following building error with Android P:

external/drm_hwcomposer/utils/hwcutils.cpp:77:31: error: assigning to 'int' from incompatible type 'android::hardware::graphics::mapper::V2_0::Error'
  ret = gm.getGrallocMapper().importBuffer(handle,
        ~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~
1 error generated.

Fixes: add24cb7 ("drm_hwcomposer: stop using pixel_stride")

+++ utils/hwcutils.cpp
@@ -76,7 +76,7 @@ int DrmHwcNativeHandle::CopyBufferHandle(buffer_handle_t handle) {
   GraphicBufferMapper &gm(GraphicBufferMapper::get());
   int ret;
 
-  ret = gm.getGrallocMapper().importBuffer(handle,
+  ret = (int) gm.getGrallocMapper().importBuffer(handle,
                                            const_cast<buffer_handle_t *>(
                                                &handle_copy));
